More Re-Writing!
*- OPs / Permission bypass added.*
Disabled by default, setting to true in config will allow it to be modified.

*- Default Message changing added.*
This can be changed via config or command.

*- /ab msg <msg> added.*
This will change the message a player sees.

*- /ab msg-reset added.*
This will reset the message back to default.

*-/ab reload added*
This will reload the configuration.

*- Removed logging temporarily*
Planning on re-modifying the logging server to be more stable down the line.

*Future*
- Specific Region Messages.

:tada: